Improve heading direction detection
Instead of checking for +1 or -1 from the initialHeading, use a circular
comparison function to check if the previous, initial, and current
heading are ordered clockwise or counterclockwise.

This now accounts for rapid changes in rotation, as well as large swings
across the initialHeading.

In addition, this will not reset the current rotation direction when
reaching the initialHeading, which allows us to check if the reset
should trigger an increment/decrement of the turn counter.