Implement better direction flow tracking
Given two indexes, a, b, a -> b could be the same, clockwise,
counter-clockwise, or directly opposite (indeterminate).

This comparison is made by checking if index b is on the left **half**
or right **half** of the circular structure.

However, with three indexes, if any two indexes are more than 1/3 radix
apart, the direction is also indeterminate.

For example, given radix of 20, and indexes a, b, c:
* a = 9, b = 0, c = 11 implies two possible orders:
  * a -> c -> b: due to closeness
  * a -> b -> c: if given large swings in index

Previously, by comparing a, b and b, c against each other, this context
was lost, and bugs were introduced when a, c were close to being
opposite b.

Therefore, a `flow` function was written to track the flow of 3 indexes
on a circular structure.